isolate stratum_apps::coinbase_output_constraints
