+++ miner-apps/jd-client/src/lib/channel_manager/mod.rs
@@ -9,14 +9,12 @@ use std::{
 
 use async_channel::{Receiver, Sender};
 use stratum_apps::{
+    coinbase_output_constraints::coinbase_output_constraints_message,
     custom_mutex::Mutex,
     key_utils::{Secp256k1PublicKey, Secp256k1SecretKey},
     network_helpers::noise_stream::NoiseTcpStream,
     stratum_core::{
-        bitcoin::{
-            absolute::LockTime, transaction::Version, OutPoint, ScriptBuf, Sequence, Target,
-            Transaction, TxIn, TxOut, Witness,
-        },
+        bitcoin::{Target, TxOut},
         channels_sv2::{
             client::extended::ExtendedChannel,
             server::{
@@ -43,9 +41,7 @@ use stratum_apps::{
         },
         noise_sv2::Responder,
         parsers_sv2::{AnyMessage, JobDeclaration, Mining, TemplateDistribution, Tlv},
-        template_distribution_sv2::{
-            CoinbaseOutputConstraints, NewTemplate, SetNewPrevHash as SetNewPrevHashTdp,
-        },
+        template_distribution_sv2::{NewTemplate, SetNewPrevHash as SetNewPrevHashTdp},
     },
     task_manager::TaskManager,
     utils::{
@@ -1091,41 +1087,11 @@ impl ChannelManager {
         &self,
         coinbase_outputs: Vec<TxOut>,
     ) -> Result<(), JDCError> {
-        // calculate the max coinbase output size for CoinbaseOutputConstraints
-        let max_size: u32 = coinbase_outputs.iter().map(|o| o.size() as u32).sum();
-        tracing::debug!(
-            max_size,
-            outputs_count = coinbase_outputs.len(),
-            "Calculated max coinbase output size"
-        );
-
-        // this is used to calculate the sigops of the coinbase outputs
-        // for CoinbaseOutputConstraints
-        let dummy_coinbase = Transaction {
-            version: Version::TWO,
-            lock_time: LockTime::ZERO,
-            input: vec![TxIn {
-                previous_output: OutPoint::null(),
-                script_sig: ScriptBuf::new(),
-                sequence: Sequence::MAX,
-                witness: Witness::from(vec![vec![0; 32]]),
-            }],
-            output: coinbase_outputs,
-        };
-
-        let max_sigops = dummy_coinbase.total_sigop_cost(|_| None) as u16;
-        tracing::debug!(max_sigops, "Calculated max sigops for coinbase");
-
-        let coinbase_output_constraints = CoinbaseOutputConstraints {
-            coinbase_output_max_additional_size: max_size,
-            coinbase_output_max_additional_sigops: max_sigops,
-        };
+        let msg = coinbase_output_constraints_message(coinbase_outputs);
 
         self.channel_manager_channel
             .tp_sender
-            .send(TemplateDistribution::CoinbaseOutputConstraints(
-                coinbase_output_constraints,
-            ))
+            .send(TemplateDistribution::CoinbaseOutputConstraints(msg))
             .await
             .map_err(|e| {
                 error!(error = ?e, "Failed to send CoinbaseOutputConstraints message to TP");
