+++ stratum-apps/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "stratum-apps"
-version = "0.1.0"
+version = "0.2.0"
 authors = ["The Stratum V2 Developers"]
 edition = "2021"
 readme = "README.md"
